Reduce docker image by binary file size
To reduce the overall docker image size I have compressed the binary
size with the help of upx. I have not executed `--brute` as this results
in an unusable binary. The image size gets decreased from 50MB down to
20MB.

Beside that I'm creating an nsswitch.conf file to make sure the Go
binary properly handles name resolution, for further reference to that
you can read up on
https://giedrius.blog/2018/11/22/etc-nsswitch-conf-and-etc-hosts-woes-with-the-alpine-and-others-docker-image-and-golang/.