[WIP] Add ARM64 compatibility to containers (#1826)
* Build AMD64 and ARM64 Docker image

By using Docker's `docker buildx` command we are able to build Docker
images for multiple platforms at the same time. It also supports
multiple tags and pushing. We use this to push all the platform specific
Docker images to the Docker registry with a single manifest.

* Build AMD64 and ARM64 micro/cells Docker image

* Build AMD64 and ARM64 Docker images in CI

I'm opting to move the Docker GitHub Action workflow from
elgohr/Publish-Docker-Github-Action to docker/build-push-action as the
latter supports [multi-platform builds][1] where the former
[doesn't][2]. This allows us to build AMD64 and ARM64 Docker images in
the pipeline. The image tags remain the same as before, e.g. `latest`
and a [timestamp][git commit short hash] tag.

[1]: https://github.com/docker/build-push-action/blob/master/docs/advanced/multi-platform.md
[2]: https://github.com/elgohr/Publish-Docker-Github-Action/issues/106

* Fix building micro/cells in CI

* Fetch tags in CI before building

This is required to build the `micro` binary with the proper `GitTag`
variable. Without this, the full Git commit hash would be used, and is
not a viable program version, which would cause `micro server` to panic.